+++ tools/bt_time.cpp
@@ -580,7 +580,7 @@ int cpp_main(int argc, char * argv[])
     return 1;
   }
 
-  if (cache_sz == (-2))
+  if (cache_sz == static_cast<std::size_t>(-2))
     cache_sz = btree::max_cache_default(common_flags, 0);
 
   cout.imbue(std::locale(std::locale(), new thousands_separator));
